+++ src/guillo-gmi/components/panel/permissions_prinperm.js
@@ -22,15 +22,11 @@ export function PermissionPrinperm({
     error: undefined,
   })
 
-  const getMultiples = (field, setter) => (ev) => {
-    let values = []
-    for (let i = 0; i < ev.target.selectedOptions.length; i++) {
-      values = values.concat([ev.target.selectedOptions[i].value])
-    }
+  const getMultiples = (field, setter) => (values) => {
     setter({ [field]: values })
   }
 
-  const savePermission = async (ev) => {
+  const savePermission = async () => {
     if (!state.principal || !state.setting || state.permission.length === 0) {
       setState({ error: 'Invalid form' })
       return
@@ -61,7 +57,7 @@ export function PermissionPrinperm({
         <Select
           appendDefault
           options={principals}
-          onChange={(ev) => setState({ principal: ev.target.value })}
+          onChange={(value) => setState({ principal: value })}
           dataTest="selectPrincipalTest"
         />
       </div>
@@ -80,7 +76,7 @@ export function PermissionPrinperm({
         <Select
           appendDefault
           options={operations}
-          onChange={(ev) => setState({ setting: ev.target.value })}
+          onChange={(value) => setState({ setting: value })}
           dataTest="operationPermissionsTest"
         />
       </div>